requirements.txt
# --- CORE INFRASTRUCTURE ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
psutil>=5.9.0
rich>=13.0.0
httpx>=0.26.0
websockets>=12.0
watchdog>=4.0.0       
python-multipart      
validators>=0.22

# --- DATA & FORMATTING ---
pydantic>=2.5.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0

# --- VOCAL CORE (UNCHAINED) ---
edge-tts>=6.1.9
pydub>=0.25.1
ffmpeg-python>=0.2.0

# --- NVIDIA NEMOTRON LOCAL CORE ---
transformers>=4.44.0
torch>=2.2.0
accelerate>=0.27.0
huggingface_hub>=0.20.0
bitsandbytes>=0.42.0
einops>=0.7.0

# --- AI ORCHESTRATION ---
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.0.1
langchain-openai>=0.0.5
langgraph>=0.0.15
tiktoken>=0.5.0
groq>=0.3.0

# --- VISION & BROWSER ---
playwright>=1.40.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0
duckduckgo-search>=6.2.1

# --- FINANCE, COMMERCE & MONETIZATION ---
yfinance>=0.2.30
scikit-learn>=1.3.0
stripe>=7.0.0
shopifyapi>=12.0.0

# --- SOCIAL MEDIA & OUTREACH ---
facebook-sdk>=3.1.0
python-linkedin-v2>=0.7.0

# --- MEMORY & GRAPH ---
chromadb>=0.4.20
fastembed>=0.2.0
networkx>=3.1

# --- INDUSTRIAL CREATIVE & DOCS ---
reportlab>=4.0.0
python-pptx>=0.6.23
replicate>=0.22.0
qrcode>=7.4
markdown>=3.5
textblob>=0.17.1

# --- UTILITIES & CLOUD ---
PyGithub>=2.1.1
isort>=5.13.0

# --- COMPATIBILITY FIXES ---
# Prevents the "Numpy 2.0" crash with LangChain on Windows
numpy<2.0.0